Fixed merged tests
