+++ dev/src/__tests__/TestHelpers.ts
@@ -1,4 +1,9 @@
+import { createOvermindMock } from "overmind"
 import { Browser, Builder, IRectangle, ThenableWebDriver } from "selenium-webdriver"
+import { MockGrpcManager } from "../MockGRPCManager"
+import { config } from "../overmind"
+import { ReviewState } from "../overmind/namespaces/review/state"
+import { State } from "../overmind/state"
 
 export const isOverlapping = (rect: IRectangle, rect2: IRectangle) => {
     return (rect.x < rect2.x + rect2.width &&
@@ -60,4 +65,19 @@ export const setupDrivers = (path?: string): ThenableWebDriver[] => {
     })
 
     return drivers
+}
+
+// initializeOvermind creates a mock overmind instance with the given state.
+// The returned overmind instance also has the grpcMan set to a MockGrpcManager
+// NOTE: Directly setting derived values in the state is not supported.
+export const initializeOvermind = (state: Partial<State>, reviewState?: Partial<ReviewState>, userID?: number) => {
+    const overmind = createOvermindMock(config, {
+        grpcMan: userID ? new MockGrpcManager(userID) : new MockGrpcManager(),
+    }, initialState => {
+        Object.assign(initialState, state)
+        if (reviewState) {
+            Object.assign(initialState.review, reviewState)
+        }
+    })
+    return overmind
 }
