+++ dev/src/__tests__/updateAdmin.test.tsx
@@ -6,32 +6,32 @@ import { config } from "../overmind"
 
 describe("Correct permission status should be set", () => {
     it("If user is not admin, promote to admin", () => {
-        const user = new User().setId(1).setName("Test User").setIsadmin(false)
+        const user = new User().setId(1).setName("Test User").setIsadmin(false).toObject()
         const mockedOvermind = createOvermindMock(config, (state) => {
             state.allUsers = [user]
         })
         window.confirm = jest.fn(() => true)
         updateAdmin(mockedOvermind, user)
-        expect(user.getIsadmin()).toBe(true)
+        expect(user.isadmin).toBe(true)
     })
 
     it("If user is admin, demote user", () => {
-        const user2 = new User().setId(2).setName("Test User2").setIsadmin(true)
+        const user2 = new User().setId(2).setName("Test User2").setIsadmin(true).toObject()
         const mockedOvermind2 = createOvermindMock(config, (state) => {
             state.allUsers = [user2]
         })
         window.confirm = jest.fn(() => true)
         updateAdmin(mockedOvermind2, user2)
-        expect(user2.getIsadmin()).toBe(false)
+        expect(user2.isadmin).toBe(false)
     })
 
     it("If user does not confirm, do not make any changes", () => {
-        const user3 = new User().setId(3).setName("Test User3").setIsadmin(true)
+        const user3 = new User().setId(3).setName("Test User3").setIsadmin(true).toObject()
         const mockedOvermind2 = createOvermindMock(config, (state) => {
             state.allUsers = [user3]
         })
         window.confirm = jest.fn(() => false)
         updateAdmin(mockedOvermind2, user3)
-        expect(user3.getIsadmin()).toBe(true)
+        expect(user3.isadmin).toBe(true)
     })
 })
